Reorganize override folder structure.
Regular monkey patching was implemented by a [previous commit](https://github.com/solidusio/solidus-demo/pull/94/commits/baa478c608feb8bd509ff78fc40aa7ce671619c7).
In response, the file structure needed to be reorganized and few files
renamed to be recognized by the application.
