+++ app/monkey_patches/spree/taxon_monkey_patch.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Spree
+  module CustomQuery
+    def featured(limit = 3)
+      all_products.featured.order('RANDOM()').limit(limit)
+    end
+
+    Taxon.prepend(CustomQuery)
+  end
+end
