--- app/overrides/solidus_demo/spree/product/custom_scope.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module SolidusDemo
-  module Spree
-    module Product
-      module CustomScope
-        def self.prepended(base)
-          base.scope :featured, -> { where(featured: true) }
-        end
-
-        ::Spree::Product.prepend self
-      end
-    end
-  end
-end
